+++ asciimatics/widgets.py
@@ -2816,14 +2816,19 @@ class FileBrowser(MultiColumnListBox):
     A FileBrowser is a widget for finding a file on the local disk.
     """
 
-    def __init__(self, height, root, name=None, on_select=None, on_change=None):
-        """
+    def __init__(self, height, root, name=None, on_select=None, on_change=None, file_filter=None):
+        r"""
         :param height: The desired height for this widget.
         :param root: The starting root directory to display in the widget.
         :param name: The name of this widget.
         :param on_select: Optional function that gets called when user selects a file (by pressing
             enter or double-clicking).
         :param on_change: Optional function that gets called on any movement of the selection.
+        :param file_filter: Optional RegEx string that can be passed in to filter the files to be displayed.
+
+        Most people will want to use a filter to finx files with a particular extension.  In this case,
+        you must use a regex that matches to the end of the line - e.g. use ".*\.txt$" to find files ending
+        with ".txt".  This ensures that you don't accidentally pick up files containing the filter.
         """
         super(FileBrowser, self).__init__(
             height,
@@ -2841,6 +2846,7 @@ class FileBrowser(MultiColumnListBox):
         self._root = root
         self._in_update = False
         self._initialized = False
+        self._file_filter = None if file_filter is None else re.compile(file_filter)
 
     def update(self, frame_no):
         # Defer initial population until we first display the widget in order to avoid race
@@ -2920,6 +2926,9 @@ class FileBrowser(MultiColumnListBox):
                     name = "|-+ {} -> {}".format(my_file, real_path)
                 else:
                     name = "|-+ {}".format(my_file)
+            elif self._file_filter and not self._file_filter.match(my_file):
+                # Skip files that don't match the filter (if present)
+                continue
             elif os.path.islink(full_path):
                 # Check if link target exists and if it does, show statistics of the
                 # linked file, otherwise just display the link